Fix PS4 shader compile errors 2

+++ crest/Assets/Crest/Crest/Shaders/OceanEmission.hlsl
@@ -76,7 +76,7 @@ half3 ScatterColour(
 		uint slice0, slice1; float lodAlpha;
 		PosToSliceIndices(samplePoint, _InstanceData.x, minSliceIndex, slice0, slice1, lodAlpha);
 
-		float2 shadowSoftHard = 0.0;
+		half2 shadowSoftHard = 0.0;
 		// TODO - fix data type of slice index in WorldToUV - #343
 		SampleShadow(_LD_TexArray_Shadow, WorldToUV(samplePoint, slice0), 1.0 - lodAlpha, shadowSoftHard);
 		SampleShadow(_LD_TexArray_Shadow, WorldToUV(samplePoint, slice1), lodAlpha, shadowSoftHard);
