+++ crest/Assets/Crest/Crest/Shaders/Resources/ShapeCombine.shader
@@ -20,7 +20,7 @@ Shader "Hidden/Crest/Simulation/Combine Animated Wave LODs"
 			CGPROGRAM
 			#pragma vertex Vert
 			#pragma fragment Frag
-			
+
 			#pragma multi_compile __ _DYNAMIC_WAVE_SIM_ON
 			#pragma multi_compile __ _FLOW_ON
 
@@ -74,10 +74,10 @@ Shader "Hidden/Crest/Simulation/Combine Animated Wave LODs"
 				const float3 uv_nextLod = WorldToUV_BiggerLod(worldPosXZ);
 
 				float3 result = 0.0;
-				float sss = 0.0;
+				half sss = 0.0;
 
 #if _FLOW_ON
-				float2 flow = 0.0;
+				half2 flow = 0.0;
 				SampleFlow(_LD_TexArray_Flow, uv_thisLod, 1.0, flow);
 
 				float2 offsets, weights;
@@ -173,7 +173,7 @@ Shader "Hidden/Crest/Simulation/Combine Animated Wave LODs"
 				o.uv = GetFullScreenTriangleTexCoord(input.VertexID);
 				return o;
 			}
-			
+
 			half4 Frag(Varyings input) : SV_Target
 			{
 				return _CombineBuffer.SampleLevel(LODData_point_clamp_sampler, input.uv, 0.0);
